add shaders for opengl rendering and have a static example working
The static example works, but we really need to make better use of
VAO's. Only when I reset the VAO, I can draw something. A lot of the
configuration is saved in the VAO, which makes using the VAO save us a
lot of trouble of passing stuff around.

Scene composition is hella hacky right now.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>